backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Werkzeug==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
